已完成 1.个人模块：创建简历、更新简历、查询简历 2.我要招聘模块：根据recruitid 查看招聘消息、查看正在进行的招聘消息 、查看已失效的招聘信息、改变招聘的状态 、查看报名列表 、查看被选中的报名列表 、选择某个应聘者 3.我要应聘：应聘的人查看招聘列表 、查看某条招聘、申请应聘、查看申请列表、取消申请、检查是否已经选择

+++ src/main/java/com/trio/breakFast/model/Tac_atoocomment.java
@@ -49,7 +49,7 @@ public class Tac_atoocomment implements Serializable{
     private String aTOoComment;//应聘者对招聘者评价
 
     @Column(name = "aTOoPoint")
-    private Integer aTOoPoint;//应聘者对招聘者评分
+    private Float aTOoPoint;//应聘者对招聘者评分
 
     @Column(name = "aTOotime")
     private String aTOotime;//应聘者对招聘者评价发布时间
